refactor: Favorite 엔티티와 리포지토리를 Account 도메인으로 이동
- Favorite는 계정별 선호 챔피언 정보로 Account와 강한 연관관계
  - 선호 챔피언은 플레이어의 게임 스타일과 숙련도를 나타내는 계정의 프로필 성격을 띔
